% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_violin.R
\name{add_violin}
\alias{add_violin}
\title{Add violins to a plot}
\usage{
add_violin(P = NULL, data, xvar_name, yvar_name, nlevels = 3)
}
\arguments{
\item{P}{The plot so far}

\item{data}{The data to be used (same as in P)}

\item{yvar_name}{the name of the y variable}

\item{xvar}{the x variable values}
}
\description{
If there is a numerical explanatory variable, break it into three divisions.
Otherwise, use the categorical explanatory variable directly.
}
