% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulers.R
\name{add_y_ruler}
\alias{add_y_ruler}
\title{Add rulers and violins to plots}
\usage{
add_y_ruler(P, x_range, ruler = NULL)
}
\arguments{
\item{P}{The plot so far}

\item{x_range}{(xmin, xmax) range of the x variable (for positioning the ruler)}

\item{ruler}{the shiny input corresponding to the ruler, e.g. `input$yruler``}
}
\description{
Add rulers and violins to plots
}
