% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_info.R
\name{get_app_info}
\alias{get_app_info}
\title{Get a listing of the Little Apps including where they are deployed}
\usage{
get_app_info(pattern = NULL)
}
\arguments{
\item{pattern}{an (optional) regex to identify specific apps

Returns a data frame giving the names of the Little Apps and the URL where they are deployed.
If \code{pattern} is given, return the info for just the corresponding little apps.}
}
\description{
Get a listing of the Little Apps including where they are deployed
}
\examples{
get_app_info("jitter")
}
