% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_ui.R
\name{LA_data_source}
\alias{LA_data_source}
\title{Basic control blocks for Little Apps}
\usage{
LA_data_source(width = 12, collapsed = FALSE, covariate = TRUE,
  facet = FALSE)
}
\arguments{
\item{width}{1-12 twitter bootstrap width of control}

\item{collapsed}{Logical indicating whether to display box as collapsed initially}
}
\description{
Basic control blocks for Little Apps
}
