% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{LA_point_alpha}
\alias{LA_point_alpha}
\title{Choose a suitable alpha level for the plot}
\usage{
LA_point_alpha(n)
}
\arguments{
\item{the}{number of dots in the plot (that is, the sample size)}
}
\description{
Based on the number of dots in the plot
}
