% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_run.R
\name{LA_run}
\alias{LA_run}
\title{Run a Little App}
\usage{
LA_run(name, where = c("browser", "viewer"))
}
\arguments{
\item{app}{character string  naming the Little App. Can also be
a regex that uniquely identifies the appl}
}
\description{
Run a Little App
}
