% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{LA_selectNumeric}
\alias{LA_selectNumeric}
\title{Selectors for variables}
\usage{
LA_selectNumeric(none = FALSE)
}
\description{
You may wish to restrict the selection of variables to a specific type,
e.g. categorical or perhaps categorical with fewer than n levels. These
functions create such selectors, which are then passed to the standard reactives.
See \code{LA_var_types} for the structure on which these selectors operate
}
