% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_for_app.R
\name{LA_template_for_app}
\alias{LA_template_for_app}
\title{Bring up a template for a new little app in the editor}
\usage{
LA_template_for_app()
}
\description{
Bring up a template for a new little app in the editor
}
