% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_type.R
\name{LA_var_types}
\alias{LA_var_types}
\title{Compute a short description of the type of a variable}
\usage{
LA_var_types(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a data frame with four fields
\itemize{
\item vname the name of the variable
\item numeric logical saying if the var is numeric
\item class the variable's class
\item n_levels how many unique levels (whether numeric or categorical)
}
}
\description{
Compute a short description of the type of a variable
}
