% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{make_model_plots}
\alias{make_model_plots}
\title{Make the basic plots}
\usage{
make_model_plots(formula, data, show_model = TRUE, sd = TRUE,
  R2 = FALSE, F = FALSE)
}
\arguments{
\item{formula}{of form A ~ B or A ~ B + C. You can also
use,  say, ns(B,2) or * instead of +. But at most two
explanatory variables.}

\item{data}{a data frame with the sample of data to display}

\item{show_model}{flag to draw the model}

\item{sd}{logical flag to show the plus-or-minus 1 sd in  the side plot}

\item{R2}{logical flag to show the R-squared in the side plot}

\item{F}{logical flag to show the F value in the side plot}
}
\value{
A list with the two plots (P1 and P2) and
the fitted model
}
\description{
A basic plot shows two or three variables:
\itemize{
\item a response
\item an explanatory variable
\item a covariate (optional)
Constructs two graphics, one showing the data
and (optionally) the fitted model, the other showing
showing the side plot comparing the distribution
of raw to  model values
}
}
\examples{
make_model_plots(wage ~ splines::ns(educ, 2)*sex,  data = mosaicData::CPS85)
}
