% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHANES2.R
\docType{data}
\name{NHANES2}
\alias{NHANES2}
\title{National Health and Nutrition Evaluation Survey data}
\format{The variables are
\itemize{
\item \code{sex}: Sex of study participant	coded as male or female
\item \code{age}: Age in years at screening of study participant. Note: Subjects 80 years or older were recorded as 80.
\item \code{marital}: Marital status of study participant. Reported for participants aged 20 years or older. One of Married, Widowed, Divorced, Separated, NeverMarried, or LivePartner (living with partner).
\item \code{ever_married }: Was the participant ever married
\item \code{race}: Reported race of study participant: Mexican, Hispanic, White, Black, or Other.
\item \code{education}: Educational level of study participant Reported for participants aged 20 years or older. One of 8thGrade, 9-11thGrade, HighSchool, SomeCollege, or CollegeGrad.
\item \code{income}: Total annual gross income for the household in US dollars. The levels are categorical: the number refers to the middle income in a category.
\item \code{income_poverty}: The ratio of income in dollars to the poverty level of income
\item \code{home_rooms}: How many rooms are in home of study participant (counting kitchen but not bathroom). 13 rooms = 13 or more rooms.
\item \code{home_own}: One of Home, Rent, or Other indicating whether the home of study participant or someone in their family is owned, rented or occupied by some other arrangement
\item \code{home_type}:` Does the participant own or rent their home
\item \code{bmi}: BMI
\item \code{bmi_who}: Body mass index category. Reported for participants aged 2 years or older. One of 12.0_18.4, 18.5_24.9, 25.0_29.9, or 30.0_plus.
\item \code{pulse}: 60 second pulse rate
\item \code{systolic}: Systolic blood pressure reading, following the procedure outlined for BPXSAR.
\item \code{diastoic}: Diastolic blood pressure reading
\item \code{testosterone}: Testerone total (ng/dL). Reported for participants aged 6 years or older.
\item \code{direct_chol}: Direct HDL cholesterol in mmol/L. Reported for participants aged 6 years or older
\item \code{tot_chol}: Total HDL cholesterol in mmol/L. Reported for participants aged 6 years or older.
\item \code{weight}: Weight in kg
\item \code{height}: Height in cm
\item \code{height_adults}: Same as \code{height}, but \code{NA} for people under 18 years.
\item \code{weight_adults}: similar to \code{height_adults} but for \code{weight}
\item \code{bmi_adults}: similar to \code{height_adults} but for \code{bmi}
\item \code{length}: Recumbent length in cm. Reported for participants aged 0 - 3 years.
\item \code{urine_vol}: Urine volume in mL – second test. Reported for participants aged 6 years or older.
\item \code{urine_flow}: Urine flow rate (urine volume/time since last urination) in mL/min – second test. Reported for participants aged 6 years or older.
\item \code{diabetes}: Study participant told by a doctor or health professional that they have diabetes. Reported for participants aged 1 year or older as Yes or No.
\item \code{diabetes_age}: Age of study participant when first told they had diabetes. Reported for participants aged 1 year or older.
\item \code{health_general}: Self-reported rating of participant's health in general Reported for participants aged 12 years or older. One of Excellent, Vgood, Good, Fair, or Poor.
\item \code{phys_health_bad_days}: Self-reported number of days participant's physical health was not good out of the past 30 days. Reported for participants aged 12 years or older.
\item \code{mental_health_bad_days}: Self-reported number of days participant's mental health was not good out of the past 30 days. Reported for participants aged 12 years or older.
\item \code{little_interest}: Self-reported number of days where participant had little interest in doing things. Reported for participants aged 18 years or older. One of None, Several, Majority (more than half the days), or AlmostAll.
\item \code{days_with_no_interest}: Does the participant have more than the occasional day when they have little  interest in doing things.
\item \code{depressed}: Self-reported number of days where participant felt down, depressed or hopeless. Reported for participants aged 18 years or older. One of None, Several, Majority (more than half the days), or AlmostAll.
\item \code{n_pregnancies}: How many times participant has been pregnant. Reported for female participants aged 20 years or older.
\item \code{n_babies}: How many of participants deliveries resulted in live births.
\item \code{age_first_baby}: Age of participant at time of first live birth. 14 years or under = 14, 45 years or older = 45. Reported for female participants aged 20 years or older.
\item \code{pregant_currently}: Pregnancy status at the time of the health examination was ascertained for females 8-59 years of age. Due to disclosure risks pregnancy status was only released for women 20-44 years of age.
\item \code{sleep_hours}: Self-reported number of hours study participant usually gets at night on weekdays or workdays. Reported for participants aged 16 years and older.
\item \code{sleep_trouble}: Participant has told a doctor or other health professional that they had trouble sleeping. Reported for participants aged 16 years and older. Coded as Yes or No.
\item \code{physically_active}: Participant does moderate or vigorous-intensity sports, fitness or recreational activities (Yes or No). Reported for participants 12 years or older.
\item \code{active_days}: Number of days in a typical week that participant does moderate or vigorous-intensity activity. Reported for participants 12 years or older.
\item \code{tv_hours}: Number of hours per day on average participant watched TV over the past 30 days. Reported for participants 2 years or older. One of 0_to_1hr, 1_hr, 2_hr, 3_hr, 4_hr, More_4_hr.
\item \code{comp_hours}: Number of hours per day on average participant used a computer or gaming device over the past 30 days. Reported for participants 2 years or older. One of 0_hrs, 0_to_1hr, 1_hr, 2_hr, 3_hr, 4_hr, More_4_hr.
\item \code{alcohol_drinks}: Average number of drinks consumed on days that participant drank alcoholic beverages. Reported for participants aged 18 years or older.
\item \code{alcohol_days}: Estimated number of days over the past year that participant drank alcoholic beverages. Reported for participants aged 18 years or older.
\item \code{smoke}: Study participant currently smokes cigarettes regularly. Reported for participants aged 20 years or older as Yes or No, provided they answered Yes to having somked 100 or more cigarettes in their life time. All subjects who have not smoked 100 or more cigarettes are listed as NA here.
\item \code{smoke100}: Study participant has smoked at least 100 cigarettes in their entire life. Reported for participants aged 20 years or older as Yes or No.
\item \code{smoke_age}: Age study participant first started to smoke cigarettes fairly regularly. Reported for participants aged 20 years or older.
\item \code{marijuana}: Participant has tried marijuana. Reported for participants aged 18 to 59 years as Yes or No.
\item \code{marijuana_age}: Age participant first tried marijuana. Reported for participants aged 18 to 59 years.
\item \code{marijuana_regular}: Participant has been/is a regular marijuana user (used at least once a month for a year). Reported for participants aged 18 to 59 years as Yes or No.
\item \code{age_sex}: Age of participant when had sex for the first time. Reported for participants aged 18 to 69 years.
\item \code{orientation_sex}: Participant's self-described sexual orientation. Reported for participants aged 18 to 59 years. One of Heterosexual, Homosexual, Bisexual.
}}
\source{
See \code{NHANES::NHANES}
}
\usage{
data(NHANES2)
}
\description{
National Health and Nutrition Evaluation Survey data
}
\details{
This is survey data collected by the US National Center for Health
Statistics (NCHS) which has conducted a series of health and nutrition surveys
since the early 1960's. Since 1999 approximately 5,000 individuals of all ages
are interviewed in their homes every year and complete the health examination
component of the survey. The health examination is conducted in a mobile
examination centre (MEC). This data.frame object is derived from \code{NHANES::NHANES}.
It simplifies some variable names and codings to make it more suitable for interactive use.
}
\keyword{datasets}
