% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_helpers.R
\docType{data}
\name{no_x_axis}
\alias{no_x_axis}
\title{Graphics helpers}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 3.}
\usage{
no_x_axis
}
\description{
Graphics helpers
}
\keyword{datasets}
