% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_plot.R
\name{one_sample_t_plot}
\alias{one_sample_t_plot}
\title{One-sample t test plot}
\usage{
one_sample_t_plot(formula, data, level = 0.95, show_mean = TRUE,
  show_ci = TRUE, null_hypothesis = 0, y_range = NULL)
}
\description{
One-sample t test plot
}
