% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density_y}
\alias{plot_density_y}
\title{Plots for striped densities
A Little App graphic dividing a density up into labeled groups.}
\usage{
plot_density_y(yvar = NULL, xvar = "all", violin = FALSE,
  show_normal = FALSE, breaks = c(-3, -2, 2, 3), qbreaks = NULL,
  labels = c("rare_left", "uncommon_left", "common", "uncommon_right",
  "rare_right"), kill_in_labels = "_.+$", show_percent = TRUE,
  vwidth = 0.35, max_x_levels = 10)
}
\arguments{
\item{violin}{Logical flag specifying whether the graph should have the
traditional density vs y frame or  be in the form of a violin plot.}

\item{show_normal}{Logical flag specifying whether to overlay a normal density}

\item{breaks}{numerical  vector of y values at which to divide the density plot}

\item{qbreaks}{numerical vector, like \code{breaks}, but giving the cumulatave probabilities at which
to divide the density plot. If \code{qbreaks} is specified, it will override \code{breaks}. That is,
\code{breaks} will be automatically set  to the quantiles defined by \code{qbreaks}.}

\item{labels}{character vector of labels for the regions defined by \code{breaks}. There
must have length 1 greater than \code{breaks}, since k numbers in \code{breaks} defines k+1 regions.
The labels must be unique: no repeats.}

\item{kill_in_labels}{a character string giving a regular expression identifying components
of each label to delete when displaying labels.  Default: \code{"_.+$"}, which deletes
any text following \code{_} in the labels. (Implication: If you want two  labels to
display identically, use \code{_} as the separator for a  unique-making suffix, e.g. \code{_left} and \code{_right}.)}

\item{show_percent}{Logical indicating whether to show the percent covered by
each of the divisions defined by \code{breaks} or \code{qbreaks}.}

\item{vwidth}{width for the violins}

\item{max_x_levels}{maximum number of levels of \code{xvar} (default 10) to throw an error indicating
that there are "too  many" breaks. This is really just a guard against accidental
specification of \code{xvar} as a numeric vector with lots and lots of levels that
would lmake the graphic confusing. As needed, override this.}

\item{yval}{numerical vector of values for which to show the density}

\item{xval}{optional vector defining the groups each \code{yval} value belongs to. Typically
a character vector or factor of the same length as \code{yval}.}
}
\description{
Plots for striped densities
A Little App graphic dividing a density up into labeled groups.
}
\examples{
plot_density_y(rnorm(10000), c(1:2))
plot_density_y(rexp(10000) * c(1, 2), xvar = c("minor", "major"),
qbreaks = c(.1, .25, .75, .9)) \%>\%
  gf_lims(x  = c(0, 6))
}
