% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmod.R
\name{qmod}
\alias{qmod}
\title{Simple model value finding function}
\usage{
qmod(data, mod_formula, testing = data, keep_all = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{mod_formula}{a formula for the model. The first of the explanatory
variables (if any) will be treated as the x variable, the rest will be the
covariates}

\item{testing}{a hold out data set for testing the model (default: use training data)}

\item{keep_all}{Flag (default: FALSE) to return all the columns in \code{data} or
just the ones named as response, explanatory variables, or covariates.}
}
\value{
A data frame with columns from the data (see \code{keep_all}), and with
\itemize{
\item \code{mod_total} the model-value column,
\item \code{from_x} the part of \code{mod_total} contributed to by the explanatory variable
\item \code{covar} the part of \code{mod_total} contributed to by the covariates (including
the intercept)
\item \code{xperp} the part of the explanatory variable perpendicular to the covariates (includ
the intercept)
}
}
\description{
Fits a model and extracts the fitted values. Some part of the
model values is assigned to the explanatory variable and the rest to the
covariates.
}
