% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_graphics.R
\name{show_bootstrap_sample_cont}
\alias{show_bootstrap_sample_cont}
\title{Graphics for bootstrapping and sampling}
\usage{
show_bootstrap_sample_cont(data, formula, ns = 3)
}
\description{
Graphics for bootstrapping and sampling
}
