% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother_plot.R
\name{smoother_plot}
\alias{smoother_plot}
\title{Make an augmented plot}
\usage{
smoother_plot(plot_formula, model_formula, data, color = "black",
  width = 0, height = 0, alpha = 0.5, trace_vert = TRUE,
  trace_horiz = TRUE, show_mod_vals = TRUE, ruler = NULL)
}
\arguments{
\item{plot_formula}{A formula of the form y ~ x}

\item{model_formula}{A formula for the model (may include nonlinearities, interactions)}

\item{data}{the data frame used to fit the model and plot the points}

\item{color}{the gf_point argument for color}

\item{width}{the gf_jitter argument for jitter width}

\item{height}{the gf_jitter argument for jitter height}

\item{alpha}{the gf_jitter argument for alpha}

\item{trace_vert}{logical. Whether to draw thin lines from the actual value to the model values}

\item{trace_horiz}{logical. Whether to draw thin lines from the model values to the bars on  the side}

\item{show_mod_vals}{logical. Whether to place a small dot at the model values}
}
\description{
A standard plot, but with bars on the side indicating spread of
actual values and model values
}
\examples{
smoother_plot(height ~ mother, height ~ splines::ns(mother, 2) * sex, Galton, color = ~ sex)


}
