% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_plot.R
\name{two_sample_t_plot}
\alias{two_sample_t_plot}
\title{Two-sample t-test plot}
\usage{
two_sample_t_plot(formula, data, level = 0.95, show_mean = TRUE,
  show_ci = TRUE, show_t = TRUE, var_equal = TRUE)
}
\description{
Two-sample t-test plot
}
