% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{LA_body}
\alias{LA_body}
\title{Builds the main panel of a little app}
\usage{
LA_body(..., plot_widget = NULL)
}
\arguments{
\item{...}{Additional \code{tabPanels()} to include in the tab box}

\item{plot_widget}{Controls how the plotting frame is set  up.
Default: \code{plotOutput("main_plot", height = "400px")}. Must come \emph{after} the
`...`` arguments.

By default, the plot widget will be set up as  a simple \code{plotOutput()} shiny object.
By using the \code{plot_widget} argument, you  can add brush  controls or choose to
substitute something entirely for the plot space in the app.}
}
\value{
The HTML which creates the dashboard body
}
\description{
You can add additional tabs via the ...
}
\examples{
LA_body()  # a default app
one <- tabPanel("A new tab", plotOutput("second_plot"), height = "400px")
two <- tabPanel("Another new tab", tableOutput("data_display"))
LA_body(one, two)

}
