% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_server.R
\name{LA_standard_reactives}
\alias{LA_standard_reactives}
\title{Basic  server elements for Little Apps}
\usage{
LA_standard_reactives(input, output, session, the_data, app_state)
}
\description{
Basic  server elements for Little Apps
}
