% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_server.R
\docType{data}
\name{LA_standard_reactives}
\alias{LA_standard_reactives}
\title{Basic  server elements for Little Apps}
\format{An object of class \code{{} of length 29.}
\usage{
LA_standard_reactives
}
\description{
Randy points out that I need to document this and have a list of the
standard reactives.
}
\keyword{datasets}
